+++ src/pages/Question.js
@@ -1,14 +1,26 @@
 import React from 'react';
 import styled from 'styled-components';
+import { useDispatch } from 'react-redux';
+import { ProgressBar } from 'components/ProgressBar';
+import { NextButton } from 'components/NextButton';
 import { CurrentQuestion } from 'components/CurrentQuestion';
 import { Devices } from '../components/MainStyles';
+import { quiz } from '../reducers/quiz';
 import halloween from '../images/halloween.jpg';
 import mobileBG from '../images/mobileBG.jpeg';
 
 export const Question = () => {
+  const dispatch = useDispatch();
+
+  const moveToNext = () => {
+    dispatch(quiz.actions.goToNextQuestion())
+  }
+
   return (
     <OuterWrapperQuestion>
       <CurrentQuestion />
+      <ProgressBar />
+      <NextButton clickAction={moveToNext} content="Next" />
     </OuterWrapperQuestion>
   )
 }
